+++ traffic_portal/test/integration/PageObjects/DeliveryServicePage.po.ts
@@ -25,6 +25,7 @@ import {browser, by, element} from 'protractor';
 interface DeliveryServices {
   Type: string;
   Name: string;
+  Tenant: string;
   validationMessage: string;
 }
 interface UpdateDeliveryService {
@@ -99,7 +100,8 @@ export class DeliveryServicePage extends BasePage {
         await this.txtDisplayName.sendKeys(deliveryservice.Name + this.randomize);
         await this.selectActive.sendKeys('Active')
         await this.selectType.sendKeys('ANY_MAP')
-        await this.selectTenant.sendKeys('- root')
+        await this.selectTenant.click();
+        await element(by.name(deliveryservice.Tenant + this.randomize)).click();
         await this.selectCDN.sendKeys('dummycdn')
         await this.txtRemapText.sendKeys('test')
         break;
@@ -109,7 +111,8 @@ export class DeliveryServicePage extends BasePage {
         await this.txtDisplayName.sendKeys(deliveryservice.Name + this.randomize);
         await this.selectActive.sendKeys('Active')
         await this.selectType.sendKeys('DNS')
-        await this.selectTenant.sendKeys('- root')
+        await this.selectTenant.click();
+        await element(by.name(deliveryservice.Tenant + this.randomize)).click();
         await this.selectCDN.sendKeys('dummycdn')
         await this.txtOrgServerURL.sendKeys('http://origin.infra.ciab.test');
         await this.txtProtocol.sendKeys('HTTP')
@@ -120,7 +123,8 @@ export class DeliveryServicePage extends BasePage {
         await this.txtDisplayName.sendKeys(deliveryservice.Name + this.randomize);
         await this.selectActive.sendKeys('Active')
         await this.selectType.sendKeys('HTTP')
-        await this.selectTenant.sendKeys('- root')
+        await this.selectTenant.click();
+        await element(by.name(deliveryservice.Tenant + this.randomize)).click();
         await this.selectCDN.sendKeys('dummycdn')
         await this.txtOrgServerURL.sendKeys('http://origin.infra.ciab.test');
         await this.txtProtocol.sendKeys('HTTP')
@@ -131,7 +135,8 @@ export class DeliveryServicePage extends BasePage {
         await this.txtDisplayName.sendKeys(deliveryservice.Name + this.randomize);
         await this.selectActive.sendKeys('Active')
         await this.selectType.sendKeys('STEERING')
-        await this.selectTenant.sendKeys('- root')
+        await this.selectTenant.click();
+        await element(by.name(deliveryservice.Tenant + this.randomize)).click();
         await this.selectCDN.sendKeys('dummycdn')
         await this.txtProtocol.sendKeys('HTTP')
         break;